+++ packages/hooks/src/useCreateOffer.ts
@@ -12,23 +12,39 @@ import useApproveCollection, {
 import useApproveCurrency, { ApproveCurrencyStep } from './useApproveCurrency'
 
 gql`
-  mutation CreateOffer($input: OfferInputBis!) {
-    createOffer(input: { offer: $input }) {
-      offer {
-        id
-        eip712Data
-        asset {
-          chainId
-          collectionAddress
-        }
-      }
+  query FetchAssetForOffer($assetId: String!) {
+    asset(id: $assetId) {
+      chainId
+      collectionAddress
     }
   }
 `
 
 gql`
-  mutation PublishOffer($offerId: UUID!, $signature: String!) {
-    publishOffer(input: { id: $offerId, signature: $signature }) {
+  mutation CreateOfferSignature($offer: OfferInputBis!) {
+    createOfferSignature(input: { offer: $offer }) {
+      eip712Data
+      timestamp
+      salt
+    }
+  }
+`
+
+gql`
+  mutation CreateOffer(
+    $offer: OfferInputBis!
+    $timestamp: Int!
+    $salt: String!
+    $signature: String!
+  ) {
+    createOffer(
+      input: {
+        offer: $offer
+        timestamp: $timestamp
+        salt: $salt
+        signature: $signature
+      }
+    ) {
       offer {
         id
       }
@@ -108,7 +124,7 @@ export default function useCreateOffer(
     }
   }, [approveCollectionActiveStep])
 
-  const createAndPublishOffer = useCallback(
+  const createOffer = useCallback(
     async ({
       type,
       quantity,
@@ -132,23 +148,11 @@ export default function useCreateOffer(
       const account = await signer.getAddress()
 
       try {
-        const { createOffer } = await sdk.CreateOffer({
-          input: {
-            type,
-            makerAddress: account.toLowerCase(),
-            assetId: assetId,
-            currencyId: currencyId,
-            quantity: quantity.toString(),
-            unitPrice: unitPrice.toString(),
-            takerAddress: takerAddress?.toLowerCase() || null,
-            auctionId: auctionId || null,
-            expiredAt: expiredAt,
-          },
-        })
-        invariant(createOffer?.offer, ErrorMessages.OFFER_CREATION_FAILED)
-        const { id: offerId, eip712Data, asset } = createOffer.offer
-
+        // approval if needed
         if (type === 'SALE') {
+          // fetch asset
+          const { asset } = await sdk.FetchAssetForOffer({ assetId })
+          invariant(asset, ErrorMessages.OFFER_CREATION_FAILED)
           // creating a new offer of type sale, approval is on the asset
           await approveCollection({
             chainId: asset.chainId,
@@ -162,19 +166,37 @@ export default function useCreateOffer(
           })
         }
 
+        // fetch offer signature
         setActiveProcess(CreateOfferStep.SIGNATURE)
+        const offer = {
+          type,
+          makerAddress: account.toLowerCase(),
+          assetId: assetId,
+          currencyId: currencyId,
+          quantity: quantity.toString(),
+          unitPrice: unitPrice.toString(),
+          takerAddress: takerAddress?.toLowerCase() || null,
+          auctionId: auctionId || null,
+          expiredAt: expiredAt,
+        }
+        const { createOfferSignature } = await sdk.CreateOfferSignature({
+          offer,
+        })
+        const { eip712Data, timestamp, salt } = createOfferSignature
+
         // sign data
         const { domain, types, message /*, primaryType */ } = eip712Data
         delete types.EIP712Domain // Hack: remove primary type from types to allow ethers detect the main type "Order" (aka: primaryType)
         const signature = await signer._signTypedData(domain, types, message)
 
-        // send signature to api
-        const { publishOffer } = await sdk.PublishOffer({
-          offerId,
+        // create offer
+        const { createOffer } = await sdk.CreateOffer({
+          offer,
           signature,
+          salt,
+          timestamp,
         })
-        invariant(publishOffer?.offer, ErrorMessages.OFFER_CREATION_FAILED)
-        return publishOffer.offer.id
+        return createOffer.offer.id
       } finally {
         setActiveProcess(CreateOfferStep.INITIAL)
       }
@@ -183,7 +205,7 @@ export default function useCreateOffer(
   )
 
   return [
-    createAndPublishOffer,
+    createOffer,
     {
       activeStep,
       transactionHash: