+++ packages/hooks/CHANGELOG.md
@@ -11,6 +11,8 @@
 
 #### Changed
 
+- Update dependency `@nft/api-graphql` to `v1.0.0-beta.10` [#129](https://github.com/liteflow-labs/liteflow-js/pull/129)
+
 #### Deprecated
 
 #### Removed