+++ packages/hooks/CHANGELOG.md
@@ -13,7 +13,8 @@
 
 - Use mutations `createCurrencyApprovalTransaction` and `createCollectionApprovalTransaction` instead of `token.approval` and `currency.approval` in `useCreateOffer` and `useAcceptOffer` hooks [#117](https://github.com/liteflow-labs/liteflow-js/pull/117)
 - Use use `createOfferFillTransaction` instead of deprecated `offer.fill` in `useAcceptOffer` hook [#118](https://github.com/liteflow-labs/liteflow-js/pull/118)
-- Update dependency `@nft/api-graphql` to `v1.0.0-beta.10` [#129](https://github.com/liteflow-labs/liteflow-js/pull/129)
+- Use new flow of offer creation using the new mutation `createOfferSignature` [#119](https://github.com/liteflow-labs/liteflow-js/pull/119)
+- Update dependency `@nft/api-graphql` to `v1.0.0-beta.11` [#129](https://github.com/liteflow-labs/liteflow-js/pull/129) [#119](https://github.com/liteflow-labs/liteflow-js/pull/119)
 
 #### Deprecated