+++ packages/hooks/CHANGELOG.md
@@ -14,6 +14,8 @@
 
 #### Fixed
 
+- Fix issue with pending ownership on lazyminted assets [#137](https://github.com/liteflow-labs/liteflow-js/pull/137)
+
 #### Security
 
 ## [v1.0.0-beta.12](https://github.com/liteflow-labs/libraries/releases/tag/v1.0.0-beta.12) - 2023-02-07