+++ packages/hooks/CHANGELOG.md
@@ -14,10 +14,14 @@
 
 #### Fixed
 
+#### Security
+
+## [v1.0.0-beta.13](https://github.com/liteflow-labs/libraries/releases/tag/v1.0.0-beta.13) - 2023-02-16
+
+#### Fixed
+
 - Fix issue with pending ownership on lazyminted assets [#137](https://github.com/liteflow-labs/liteflow-js/pull/137)
 - Fix future issues with config type and too permissive type [#138](https://github.com/liteflow-labs/liteflow-js/pull/138)
-
-#### Security
 
 ## [v1.0.0-beta.12](https://github.com/liteflow-labs/libraries/releases/tag/v1.0.0-beta.12) - 2023-02-07